requirements.txt
pyuno
pyyaml
